+++ packages/MagicFilter/example/eg_mono_v1.jl
@@ -21,3 +21,11 @@ magicgeno = formmagicgeno(genofile,pedfile);
 # cd(@__DIR__)
 # outid = "sim"
 # rm.(filter(x->occursin(outid,x), readdir()))
+
+fls = [0.005, 0.0359, 0.0669, 0.0978, 0.1288, 0.3762, 0.4072, 0.4381, 0.4691, 0.5, 0.5309, 0.5619, 0.5928, 0.6237, 0.8712, 0.9022, 0.9331, 0.9641, 0.995]
+
+fls =  [0.005, 0.5, 0.995]
+
+n1 = 0
+n2 = 6
+MagicFilter.cal_post_prob(n1,n2,fls)